requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
apscheduler==3.6.3
    # via python-telegram-bot
argparse==1.4.0
    # via -r requirements.in
certifi==2020.12.5
    # via python-telegram-bot
cffi==1.14.5
    # via cryptography
cryptography==3.4.6
    # via python-telegram-bot
decorator==4.4.2
    # via -r requirements.in
dynaconf==3.1.8
    # via -r requirements.in
pycparser==2.20
    # via cffi
python-telegram-bot==13.2
    # via -r requirements.in
pytz==2021.1
    # via
    #   apscheduler
    #   python-telegram-bot
    #   tzlocal
pyyaml==5.4.1
    # via -r requirements.in
six==1.15.0
    # via apscheduler
tornado==6.1
    # via python-telegram-bot
tzlocal==2.1
    # via apscheduler

# The following packages are considered to be unsafe in a requirements file:
# setuptools
